requirements.txt
asgiref==3.5.2
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
defusedxml==0.5.0
dj-database-url==1.0.0
Django==4.1
django-environ==0.9.0
django-filter==2.0.0
django-heroku==0.3.1
django-rest-auth==0.9.3
djangorestframework==3.11.2
gunicorn==20.1.0
idna==3.3
Markdown==3.0.1
oauthlib==2.1.0
psycopg2==2.9.3
python3-openid==3.1.0
pytz==2018.7
requests==2.28.1
requests-oauthlib==1.0.0
six==1.11.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.2.0
djoser
